refactor: moved FacilityDTOs to new package
